Upgrade Rust to 1.98.1, Node to 24 LTS, and backend/frontend dependencies (#158)
Rust 1.89.0 -> 1.98.1 and all backend crates to latest (warp stays 0.3.7 for async-graphql-warp 7). Node 18/20 -> 24 LTS; frontend deps to latest within major, build/test/lint tooling to latest major, react-router-dom 6 -> 7 for GHSA-wrjc-x8rr-h8h6. Backend serves HTTP/1.1 only (h2 0.3, RUSTSEC-2026-0258, allowlisted as unreachable), keeping remote.addr in request logs. cargo audit and npm audit both clean.

+++ backend/crates/econ-graph-services/src/services/global_analysis_service.rs
@@ -1211,7 +1211,7 @@ mod tests {
             .expect("Should find China as trade partner");
 
         assert!(china_partner.country.name.starts_with("TEST China"));
-        assert!(china_partner.trade_value_usd > BigDecimal::from(0));
+        assert!(china_partner.trade_value_usd > 0);
         assert!(china_partner.trade_intensity > 0.0);
         assert!(!china_partner.relationship_type.is_empty());
     }
